==> data-analyst/p4-data-visualization/01_data_gathering_flt.ipynb <==
import pandas as pd
import numpy as np

# Financial Cognative questions

df_cog = pd.read_csv("./data/0_raw/cy07_msu_flt_cog.csv")

df_cog.shape

# Financial math questions
fi_math_question_ids = []
with open("./data/reference/finance_math_ids.txt") as math_file:
    for line in math_file:
        fi_math_question_ids.append(line.strip())

# Financial reading questions
fi_reading_question_ids = []
with open("./data/reference/finance_reading_ids.txt") as reading_file:
    for line in reading_file:
        fi_reading_question_ids.append(line.strip())

# Financial literacy and information technology questisons
fi_it_question_ids = []
with open("./data/reference/finance_it_ids.txt") as fiit_file:
    for line in fiit_file:
        fi_it_question_ids.append(line.strip())

# drop columns not planning to use to free memory
cog_cols = fi_math_question_ids + fi_reading_question_ids + fi_it_question_ids
useful_cols = ['CNT', 'CNTRYID', 'CNTSCHID', 'CNTSTUID', 'STRATUM', 'LANGTEST_COG']
df_cog.drop(columns=df_cog.columns.difference(cog_cols + useful_cols), inplace=True)
df_cog.shape

# Given column ids as question_ids this method calculates score of a row
# valid answers given 10 point, partial answers given5 points
def calculate_score_for_columns(row, question_ids):
    student_score = 0.0
    student_answered_questions = 0

    for question in question_ids:
        score = row[question]
        if not np.isnan(score):
            score_str = str(int(score))
            student_answered_questions += 1
            if score_str == "1" or score_str.startswith("2"):
                student_score += 10
            elif score_str.startswith("1"):
                student_score += 5 # partial answer, half point
    return student_score, student_answered_questions

def calculate_test_score(row):
    m_score, m_count = calculate_score_for_columns(row, fi_math_question_ids)
    r_score, r_count = calculate_score_for_columns(row, fi_reading_question_ids)
    i_score, i_count = calculate_score_for_columns(row, fi_it_question_ids)
    return pd.Series([m_score, m_count, r_score, r_count, i_score, i_count])

# Calculate scores and answer counts
score_columns = ["fi_math_score", "fi_math_answered", "fi_reading_score", "fi_reading_answered", "fi_it_score", "fi_it_answered"]
df_cog[score_columns] = df_cog.apply(lambda row: calculate_test_score(row), axis=1)

columns_to_keep = useful_cols + score_columns

# drop unused columns
df_cog.drop(columns=df_cog.columns.difference(columns_to_keep), inplace=True)

df_cog.sample(5)

df_cog.to_csv("./data/1_processed/financial_cognitive_scores.csv",index=False)

# Financial Questionnaire

- It appears that financial questionnaire `cy07_msu_flt_qqq` is already included in student questionnaire `cy07_msu_stu_qqq` which is processed in `01_data_gathering_std`
