==> data-analyst/p4-data-visualization/01_data_gathering_sch.ipynb <==
import pandas as pd

df_sch = pd.read_csv("./data/0_raw/cy07_msu_sch_qqq.csv")

df_sch.shape

# columns I find interesting
columns_to_keep = []
with open("./data/reference/columns_to_keep_sch_qqq.txt") as text_file:
    for line in text_file:        
        columns_to_keep.append(line.strip())

df_sch.drop(columns=df_sch.columns.difference(columns_to_keep), inplace=True)
df_sch.shape

df_sch.sample(5)

df_sch.to_csv("./data/1_processed/school_questionnaire.csv",index=False)

# Teacher questionnare


df_tch = pd.read_csv("./data/0_raw/cy07_msu_tch_qqq.csv")

df_tch.shape

# columns I find interesting
columns_to_keep_tch = []
with open("./data/reference/columns_to_keep_tch_qqq.txt") as text_file:
    for line in text_file:        
        columns_to_keep_tch.append(line.strip())

df_tch.drop(columns=df_tch.columns.difference(columns_to_keep_tch), inplace=True)
df_tch.shape

df_tch.to_csv("./data/1_processed/teacher_questionnaire.csv",index=False)
