==> data-analyst/p4-data-visualization/03_data_wrangling_clean.ipynb <==
import pandas as pd
import numpy as np
import json

### Load Data

column_lookup = {}
with open("./data/reference/column_lookup.txt") as text_file:
    for line in text_file:
        item = json.loads(line)
        column_lookup[item['column']] = item['meaning']

df_student = pd.read_csv("./data/2_merged/student.csv")

df_student.shape

df_student.CNTSCHID.nunique()

df_school = pd.read_csv("./data/2_merged/school.csv")

df_teacher = pd.read_csv("./data/2_merged/teacher.csv")

### Teacher dataframe prepare for merge

# Teacher columns
for column in df_teacher.columns:
    print('{:<14} -> {}'.format(column, column_lookup[column]))

# satisfied, age, # years as teacher
df_sum_teacher = df_teacher.groupby(['CNTSCHID','CNTRYID'])[['tch_satisfied', 'TC002Q01NA', 'TC007Q01NA']].mean().reset_index()

df_sum_teacher = df_sum_teacher.rename(columns={
    "tch_satisfied": "mean_teacher_satisfied",
    "TC002Q01NA": "mean_age_of_teachers",
    "TC007Q01NA": "mean_yaer_experiance"
})

df_sum_teacher.head()

### School dataframe prepare for merge

df_school.shape

# school columns
for column in df_school.columns:
    print('{:<23} -> {}'.format(column, column_lookup[column]))

# schoolid, public/private, class size,
df_sum_school = df_school[['CNTRYID', 'CNTSCHID', 'SC013Q01TA', 'SC003Q01TA', 'education_obstacle', 'problematic_students', 'degree_score_teachers']]

df_sum_school.head()

### Student dataframe prepare for merge

# dropping OECD contry and birth month columns
df_student.drop(columns=['OECD', 'ST003D02T'], inplace=True, axis=1)
df_student.shape

# Teacher columns
for column in df_student.columns:
    print('{:<20} -> {}'.format(column, column_lookup[column]))

### Merge dataframes

df_student.shape

df_sum_school.shape

df_final = pd.merge(df_student, df_sum_school, on=["CNTSCHID","CNTRYID"], how="outer")
df_final.shape

df_final = pd.merge(df_final, df_sum_teacher, on=["CNTSCHID","CNTRYID"], how="outer")
df_final.shape

pd.get_option("display.max_columns")


pd.set_option('display.max_columns', 40)
pd.set_option('display.max_rows', 40)

df_final.sample(10)

### Explore

df_final.info()

column_lookup['PA042Q01TA']

df_final.duplicated().sum()

df_final.CNTSCHID.nunique()

- unique student id should be country + studentId

df_final.describe()

# Teacher columns
for column in df_final.columns:
    print('{:<24} -> {}'.format(column, column_lookup[column]))

df_final= df_final.rename(columns={
"CNTRYID": "country_id",
"CNT":"country_code",
"CNTSCHID": "school_id",
"CNTSTUID": "student_id",
"STRATUM": "social_level_id",
"ST001D01T": "grade",
"ST003D03T": "birth_year",
"ST004D01T": "gender",
"PA042Q01TA": "income",
"ESCS": "socio_economic_index",
"WEALTH": "wealth",
"PV1MATH": "score_math",
"PV1READ": "score_reading",
"PV1SCIE": "score_science",
"PV1RCLI": "score_find_info",
"PV1RCUN": "score_understand",
"PV1RCER": "score_evaluate",
"home_art": "home_art_score",
"bad_home_conditions": "bad_home_conditions",
"school_bad_time": "bullied_at_school",
"emotional_status": "emotional_status",
"family_hi_ed": "family_high_edu",
"quality_tch": "teacher_quality_stu",
"reading_not_easy": "reading_not_easy",
"SC013Q01TA": "school_type",
"SC003Q01TA": "class_size",
"math_score": "cust_math_score",
"math_answered": "num_math_answered",
"reading_score": "cust_reading_score",
"reading_answered": "num_reading_answered",
"science_score": "cust_science_score",
"science_answered": "num_science_answered",
})

df_final.head()

int_columns = ["country_id","school_id","student_id","grade","birth_year","home_art_score","bad_home_conditions","bullied_at_school","family_high_edu","teacher_quality_stu","reading_not_easy","class_size","education_obstacle","school_type","gender", "cust_math_score","num_math_answered","cust_reading_score","num_reading_answered","cust_science_score","num_science_answered", "problematic_students"]
type_dict = {}
for item in int_columns:
    type_dict[item] = "Int64"

df_final = df_final.astype(type_dict)

# Round test score values
round_columns = ["score_math","score_reading","score_science","score_find_info","score_understand","score_evaluate"]
df_final[round_columns] = df_final[round_columns].round(decimals=0).astype("Int64")

df_final.head()

df_final.to_csv("./data/3_cleaned/students.csv",index=False)
